requirements.txt
tensorflow==1.4.0
flask==0.12.2
werkzeug==0.12.2
pillow==4.2.1
cassandra-driver==3.15.1

